+++ apps/client/vite.config.ts
@@ -1,0 +1,41 @@
+import { defineConfig } from "vite";
+import path from "path";
+import { fileURLToPath } from "url";
+import react from "@vitejs/plugin-react-swc";
+
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
+
+export default defineConfig({
+  base: "./",
+  plugins: [react()],
+  server: {
+    port: 3000,
+    open: false,
+  },
+  resolve: {
+    alias: {
+      components: path.resolve(__dirname, "./src/components"),
+      utils: path.resolve(__dirname, "./src/utils"),
+      hooks: path.resolve(__dirname, "./src/hooks"),
+      src: path.resolve(__dirname, "./src"),
+      plugins: path.resolve(__dirname, "./src/plugins"),
+    },
+  },
+  build: {
+    target: "baseline-widely-available",
+    outDir: "build",
+    sourcemap: false,
+    rollupOptions: {
+      output: {
+        manualChunks: {
+          "app-core": ["./src/components/App.jsx"],
+        },
+      },
+    },
+  },
+  esbuild: {
+    loader: "jsx",
+    include: /src\/.*\.jsx?$/,
+    exclude: [],
+  },
+});